Fix annoying visual bug with colors
We are not like every other rom. We are the only rom to do 2 seperate
builds of MusicFX due to the fact that we refuse to let the pixel
revolution take over our ROM. You want a vanilla expirence? Then we cant
just use the same commits as everyone else. One commit we merged changed
the stock colors of musicfx to be the pixel colors. This commit should
revert that.